fix: Run formatter

+++ src/Mocks/AggregatorV3InterfaceMock.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.9;
 import "@chainlink/contracts/src/v0.8/interfaces/AggregatorV3Interface.sol";
 
 contract PriceConsumerV3Mock {
-    function getLatestPrice(address priceFeedAddress) public view returns (int, uint8) {
+    function getLatestPrice(address priceFeedAddress) public view returns (int256, uint8) {
         return (287908444994, 8);
     }
 }
